[api] neo4j refactoring

+++ BookApp.API/Dtos/CatalogItemDto.cs
@@ -7,8 +7,8 @@ namespace BookApp.API.Dtos {
     public CatalogItemDto () {
       this.Books = new List<BookItemDto> ();
     }
-    
-    public int Id { get; set; }
+
+    public long Id { get; set; }
     public string Name { get; set; }
     public bool IsPublic { get; set; }
     public int UserId { get; set; }
